+++ Source/libtvm/BinaryWriter.h
@@ -62,13 +62,11 @@ private:
 
     size_t calculateInstructionSize(void);
 
-    size_t findLabel(const str_t& name);
-
-    size_t addToStringTable(const str_t& symname);
-    size_t addToDataTable(const DataDeclaration& dt);
-
-    size_t addLinkedSymbol(const str_t& symname, const str_t& libname);
-    int    loadSharedLibrary(const str_t& lib);
+    uint64_t findLabel(const str_t& name);
+    uint64_t addToStringTable(const str_t& symname);
+    uint64_t addToDataTable(const DataDeclaration& dt);
+    uint64_t addLinkedSymbol(const str_t& symname, const str_t& libname);
+    int      loadSharedLibrary(const str_t& lib);
 
 public:
     BinaryWriter(const str_t& modpath);
