+++ Source/stdlib/stdlib.cpp
@@ -35,7 +35,7 @@ SYM_API SYM_EXPORT void __putchar(tvmregister_t regi)
 
 SYM_API SYM_EXPORT void __puts(tvmregister_t regi)
 {
-    size_t ptr = prog_get_register64(regi, 0);
+    size_t ptr = (size_t)prog_get_register64(regi, 0);
     if (ptr)
     {
         puts((char*)ptr);
