+++ Test/fcmp/Main.cpp
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <string.h>
 
-char *GetFilteredBuffer(const char *fname, int64_t &len)
+char *GetFilteredBuffer(const char *fname, size_t &len)
 {
     FILE *fp = fopen(fname, "rb");
     if (!fp)
@@ -27,7 +27,7 @@ char *GetFilteredBuffer(const char *fname, int64_t &len)
     fread(bufB, 1, len, fp);
     fclose(fp);
 
-    int64_t i, j = 0;
+    size_t i, j = 0;
     for (i = 0; i < len; ++i)
     {
         char ch = bufB[i];
@@ -57,9 +57,10 @@ int main(int argc, char **argv)
         printf("Usage: fcmp file1 file2\n");
         return 1;
     }
-    int64_t a, b;
+    size_t  a, b;
     char *  fileA = GetFilteredBuffer(argv[1], a);
     char *  fileB = GetFilteredBuffer(argv[2], b);
+
 
     int rc = 0;
     if (a != b)
